requirements.txt
numpy==1.24.4
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
librosa==0.10.1
soundfile==0.12.1
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.7.2
streamlit==1.28.0